requirements.txt
# Core RAG Framework
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.10

# LLM and Model Management
ollama==0.1.7
transformers==4.37.0
sentence-transformers==2.7.0
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
accelerate==0.26.0
bitsandbytes==0.42.0

# Vector Database
chromadb==0.4.22
numpy==1.24.4
faiss-cpu==1.7.4

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
unstructured==0.11.8
tiktoken==0.5.2
pdfplumber==0.9.0

# Text Processing
nltk==3.8.1
spacy==3.7.2
regex==2023.12.25

# Evaluation
rouge-score==0.1.2
datasets==2.16.1
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# Web Interface
gradio==4.15.0
streamlit==1.29.0
fastapi==0.108.0
uvicorn==0.25.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
loguru==0.7.2
psutil==5.9.7

# Development
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.28.0
black==23.12.1
flake8==7.0.0
pytest==7.4.4 